+++ pywce/src/services/hook_service.py
@@ -1,6 +1,7 @@
 import importlib
+import inspect
 from functools import wraps
-from typing import Callable
+from typing import Callable, Literal, Optional
 
 from pywce.src.exceptions import HookError
 from pywce.src.models import HookArg
@@ -11,6 +12,9 @@ _logger = pywce_logger(__name__)
 # Global registries for hooks
 _hook_registry = {}
 _dotted_path_registry = {}
+_function_cache = {}
+_global_pre_hooks = []
+_global_post_hooks = []
 
 class HookService:
     """
@@ -21,7 +25,6 @@ class HookService:
     Dynamically call hook functions or class methods.
     All hooks should accept a [HookArg] param and return a [HookArg] response.
     """
-
     @staticmethod
     def registry():
         return _hook_registry
@@ -47,6 +50,36 @@ class HookService:
         _logger.debug("Registered %s hook: %s", "func" if func else "path", name)
 
     @staticmethod
+    def register_global_hook(hook_dotted_path: str, hook_type: Literal["pre", "post"]):
+        """
+        Register a global pre or post hook.
+        :param hook_dotted_path: Dotted path to the hook function.
+        :param hook_type: Either "pre" or "post".
+        """
+        if hook_type == "pre":
+            _global_pre_hooks.append(hook_dotted_path)
+        elif hook_type == "post":
+            _global_post_hooks.append(hook_dotted_path)
+        else:
+            raise HookError("Invalid hook_type. Use 'pre' or 'post'.")
+
+    @staticmethod
+    def register_callable_global_hooks(pre: list[Callable], post: list[Callable]):
+        for pre_hook in pre:
+            dotted_path = f"{pre_hook.__module__}.{pre_hook.__name__}"
+
+            func = HookService.load_function_from_dotted_path(dotted_path)
+            HookService.register_hook(name=dotted_path, func=func)
+            HookService.register_global_hook(dotted_path, "pre")
+
+        for post_hook in post:
+            dotted_path = f"{post_hook.__module__}.{post_hook.__name__}"
+
+            func = HookService.load_function_from_dotted_path(dotted_path)
+            HookService.register_hook(name=dotted_path, func=func)
+            HookService.register_global_hook(dotted_path, "post")
+
+    @staticmethod
     def load_function_from_dotted_path(dotted_path: str) -> Callable:
         """
         Load a function or attribute from a given dotted path.
@@ -55,6 +88,9 @@ class HookService:
         :return: A callable function or method.
         """
         try:
+            if dotted_path in _function_cache:
+                return _function_cache[dotted_path]
+
             if not dotted_path:
                 raise ValueError("Dotted path cannot be empty.")
 
@@ -72,13 +108,14 @@ class HookService:
             if not callable(function):
                 raise ValueError(f"Resolved object '{function_name}' is not callable.")
 
+            _function_cache[dotted_path] = function
             return function
 
         except (ImportError, AttributeError, ValueError) as e:
             raise ImportError(f"Could not load function from dotted path '{dotted_path}': {e}")
 
     @staticmethod
-    def process_hook(hook_dotted_path: str, hook_arg: HookArg) -> HookArg:
+    async def _execute_hook(hook_dotted_path: str, hook_arg: HookArg) -> HookArg:
         """
         Execute a function from registry or lazy loading it.
 
@@ -101,34 +138,69 @@ class HookService:
                 hook_func = HookService.load_function_from_dotted_path(hook_dotted_path)
                 HookService.register_hook(name=hook_dotted_path, dotted_path=hook_dotted_path)
 
+            # implement class based hook
+
+            # Function-based hook
+            if inspect.iscoroutinefunction(hook_func):
+                return await hook_func(hook_arg)
+
             return hook_func(hook_arg)
 
         except Exception as e:
             _logger.error("Hook processing failure. Hook: '%s', error: %s", hook_dotted_path, str(e))
             raise HookError(f"Failed to execute hook: {hook_dotted_path}") from e
 
+    @staticmethod
+    async def process_hook(hook_dotted_path: str, hook_arg: HookArg) -> HookArg:
+        """
+        Execute a function from registry or lazy loading it.
 
-def hook(func: Callable) -> Callable:
+        :param hook_dotted_path: The dotted path to the hook function.
+        :param hook_arg: The argument to pass to the hook function.
+        :return: The result of the hook function.
+        """
+        return await HookService._execute_hook(hook_dotted_path, hook_arg)
+
+    @staticmethod
+    async def process_global_hooks(hook_type: Literal["pre", "post"], hook_arg: HookArg) -> None:
+        try:
+            hooks = _global_pre_hooks if hook_type == "pre" else _global_post_hooks
+            for pre_hook in hooks:
+                await HookService._execute_hook(pre_hook, hook_arg)
+        except HookError as e:
+            _logger.critical("Global `%s` hook processing failure, error: %s", hook_type, e.message)
+
+
+# decorator
+def hook(func: Callable, global_type: Optional[Literal["pre", "post"]] = None) -> Callable:
     """
     Decorator to register a hook function with validation.
 
     :param func: The hook function to decorate.
+    :param global_type: The type of the global hook being registered.
     :return: The wrapped function.
     """
 
-    @wraps(func)
-    def wrapper(arg: HookArg) -> HookArg:
-        if not isinstance(arg, HookArg):
-            raise HookError(f"Expected HookArg instance, got {type(arg).__name__}")
+    def decorator(inner_func: Callable) -> Callable:
+        @wraps(inner_func)
+        def wrapper(arg: HookArg) -> HookArg:
+            if not isinstance(arg, HookArg):
+                raise HookError(f"Expected HookArg instance, got {type(arg).__name__}")
+            return inner_func(arg)
 
-        _logger.debug("Invoking hook: %s", func.__name__)
-        return func(arg)
+        # Compute the full dotted path for the function
+        full_dotted_path = f"{inner_func.__module__}.{inner_func.__name__}"
 
-    # Compute the full dotted path for the function
-    full_dotted_path = f"{func.__module__}.{func.__name__}"
+        # Eagerly register the hook
+        if full_dotted_path not in _hook_registry:
+            HookService.register_hook(name=full_dotted_path, func=wrapper)
 
-    # Eagerly register the hook
-    if full_dotted_path not in _hook_registry:
-        HookService.register_hook(name=full_dotted_path, func=func)
+        if global_type:
+            HookService.register_global_hook(full_dotted_path, global_type)
 
-    return wrapper
+        return wrapper
+
+    if func:
+        return decorator(func)
+
+    return decorator